requirements-dev.txt
# Development and Testing Dependencies for AI-102
# These are required for running tests for the dashboard server with async HTTP and WebSocket support

# Core Testing Framework
pytest>=8.0.0
pytest-aiohttp>=1.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Browser Testing (Playwright)
playwright>=1.40.0

# WebSocket Support
websockets>=12.0

# Additional testing utilities
coverage>=7.0.0
